Add functionality to inventory
Items can be picked up and viewed in the inventory. All they need to do
is be a child of the "o_inventory_item" class when placed in the room,
and they can be picked up. For now, you can't remove items from your
inventory, and I've absolutely no idea if they still exist after
changing rooms. We'll find out later